Title: 바이러스 검사, Time: 371ms, Memory: 8MB, Status: Passed - CodeTree

+++ 250306/바이러스 검사/virus-detector.cpp
@@ -11,23 +11,23 @@ int main() {
     //n개의 식당 고객들의 체온을 측정하기 위해 필한 검사자 수의 최솟값
 
     int n;  cin>>n;
-    vector<int> customers(n);
+    vector<long> customers(n);
     for(int i = 0; i < n; i++)
     {
         cin>>customers[i];
     }
 
-    int checkFromLeader = 0; int checkFromMember = 0;
+    long checkFromLeader = 0; long checkFromMember = 0;
     cin>>checkFromLeader>>checkFromMember;
     for(int i = 0; i < n; i++)
     {
-        customers[i] = max(0, customers[i] - checkFromLeader);
+        customers[i] = max(0L, customers[i] - checkFromLeader);
     }
 
-    int res = 0;
+    long res = 0;
     for(int i = 0; i < n; i++)
     {
-        int memberCnt = customers[i] / checkFromMember;
+        long memberCnt = customers[i] / checkFromMember;
         if(customers[i] % checkFromMember > 0)
             ++memberCnt;
         res += memberCnt + 1;
